Move Ada news section to reusable component
